Fixed to_dbreference in mongo adapter
