+++ tests/models/test_phenotype_term.py
@@ -1,28 +1,14 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, unicode_literals
 
-from scout.models import PhenotypeTerm
+from .setup_objects import setup_phenotype_term
 
-def setup_phenotype(**kwargs):
-  """
-  Setup an Phenotype term object
-  """
-  phenotype_id = kwargs.get('hpo_id', "1234")
-  feature = kwargs.get('feature', "NOC1")
-  disease_models = kwargs.get('disease_models', ["AD"])
 
-  term = PhenotypeTerm(
-    phenotype_id=phenotype_id, 
-    feature=feature,
-    disease_models=disease_models
-  )
-  return term
-  
 def test_phenotype_term():
   """
   Test the PhenotypeTerm class
   """
-  term = setup_phenotype()
+  term = setup_phenotype_term()
   
   assert term.phenotype_id == "1234"
   assert term.feature == "NOC1"
